+++ test/eslint-plugin-e2e/test/fixtures/monorepo/package-b/helpers/helper-a/helper-a.js
@@ -1,0 +1,6 @@
+// Importing helper from same package
+import { HelperB } from "../helper-b";
+
+export const HelperA = () => {
+  return HelperB();
+};